refactor: split settings tabs into files
