+++ Midibard/UI/Settings/GeneralTab.cs
@@ -1,0 +1,167 @@
+using System.IO;
+
+using Dalamud.Interface;
+using Dalamud.Interface.ImGuiNotification;
+using Dalamud.Interface.Utility;
+
+using ImGuiNET;
+
+using MidiBard.IPC;
+using MidiBard.Managers;
+using MidiBard.Util;
+
+using static MidiBard2.Resources.Language;
+
+namespace MidiBard;
+
+public partial class PluginUI
+{
+    private void DrawGeneralSettings()
+    {
+        ImGuiGroupPanel.BeginGroupPanel(setting_group_label_general_settings);
+        {
+            if (ImGui.Checkbox(setting_label_auto_open_on_startup, ref MidiBard.config.AutoOpenOnStartup))
+            {
+                IPCHandles.SyncAllSettings();
+            }
+            ImGuiUtil.ToolTip(setting_label_auto_open_on_startup);
+
+            //-------------------
+
+            if (ImGui.Checkbox(setting_label_auto_open_when_performing, ref MidiBard.config.AutoOpenPlayerWhenPerforming))
+            {
+                IPCHandles.SyncAllSettings();
+            }
+            ImGuiUtil.ToolTip(setting_label_auto_open_when_performing);
+
+            if (ImGui.Checkbox(setting_label_auto_close_when_performing, ref MidiBard.config.AutoClosePlayerWhenPerforming))
+            {
+                IPCHandles.SyncAllSettings();
+            }
+            ImGuiUtil.ToolTip(setting_label_auto_close_when_performing);
+
+            //-------------------
+
+            if (ImGui.Checkbox(setting_label_show_now_playing_info, ref MidiBard.config.showNowPlayingInfo))
+            {
+                IPCHandles.SyncAllSettings();
+            }
+            ImGuiUtil.ToolTip(setting_label_show_now_playing_info);
+
+            //-------------------
+
+            if (ImGui.Checkbox(setting_label_hide_player_information_from_ui, ref MidiBard.config.hidePlayerInformationFromUi))
+            {
+                IPCHandles.SyncAllSettings();
+            }
+            ImGuiUtil.ToolTip(setting_label_hide_player_information_from_ui);
+
+            //-------------------
+
+            //Checkbox(Low_latency_mode, ref MidiBard.config.LowLatencyMode);
+            //ImGuiUtil.ToolTip(low_latency_mode_tooltip);
+
+            //ImGui.Checkbox(checkbox_auto_restart_listening, ref MidiBard.config.autoRestoreListening);
+            //ImGuiUtil.ToolTip(checkbox_auto_restart_listening_tooltip);
+
+            //ImGui.SameLine(ImGuiUtil.GetWindowContentRegionWidth() / 2);
+            //ImGui.Checkbox("Auto listening new device".Localize(), ref MidiBard.config.autoStartNewListening);
+            //ImGuiUtil.ToolTip("Auto start listening new midi input device when idle.".Localize());
+            //ImGuiUtil.ColorPickerButton(1000, label_theme_color, ref MidiBard.config.themeColor,
+            //    ImGuiColorEditFlags.AlphaPreview | ImGuiColorEditFlags.AlphaBar);
+            //if (ImGui.ColorEdit4("Theme color".Localize(), ref MidiBard.config.themeColor,
+            //    ImGuiColorEditFlags.AlphaPreview | ImGuiColorEditFlags.AlphaBar | ImGuiColorEditFlags.NoInputs))
+            ImGui.Spacing();
+            ImGui.Spacing();
+
+            ImGui.TextUnformatted(setting_label_theme_color);
+            ImGui.Spacing();
+            ImGui.ColorEdit4("##{setting_label_theme_color}", ref MidiBard.config.themeColor,
+                ImGuiColorEditFlags.AlphaPreview | ImGuiColorEditFlags.AlphaBar);
+
+            ImGui.SameLine();
+            if (ImGuiUtil.IconButton(FontAwesomeIcon.Undo, "##btnResetUIColor", "Reset"))
+            {
+                MidiBard.config.themeColor = Theme.Colors.Lavender;
+                IPCHandles.SyncAllSettings();
+            }
+            //-------------------
+
+            ImGui.Spacing();
+            ImGui.TextUnformatted(setting_label_played_song_highlight_color);
+            ImGui.Spacing();
+            ImGui.ColorEdit4(setting_label_played_song_highlight_color, ref MidiBard.config.playedSongColor, ImGuiColorEditFlags.NoAlpha | ImGuiColorEditFlags.NoLabel);
+            ImGui.SameLine();
+            if (ImGuiUtil.IconButton(FontAwesomeIcon.Undo, "##btnResetSongHighlightColor", "Reset"))
+            {
+                MidiBard.config.playedSongColor = Theme.Colors.Cyan;
+                IPCHandles.SyncAllSettings();
+            }
+
+            //-------------------
+
+            ImGui.Spacing();
+            ImGui.Spacing();
+            ImGui.Spacing();
+
+            if (ImGui.Combo(setting_label_select_ui_language, ref MidiBard.config.uiLang, uilangStrings,
+                    uilangStrings.Length))
+            {
+                MidiBard.ConfigureLanguage(MidiBard.GetCultureCodeString((MidiBard.CultureCode)MidiBard.config.uiLang));
+            }
+
+            //-------------------
+
+            ImGui.Spacing();
+            ImGui.Separator();
+            ImGui.Spacing();
+
+            ImGui.Text($"Default Performer Folder:");
+            ImGui.TextUnformatted(Path.ChangeExtension(MidiBard.config.defaultPerformerFolder, null).EllipsisString(70));
+
+            var btnChangeText = "Change";
+            var btnChangeSize = ImGuiHelpers.GetButtonSize(btnChangeText);
+            ImGui.SameLine(ImGui.GetWindowWidth() - 2 * ImGui.GetCursorPosX() - btnChangeSize.X);
+            if (ImGui.Button(btnChangeText))
+            {
+                RunSetDefaultPerformerFolderImGui();
+            }
+
+            ImGui.Spacing();
+
+            //-------------------
+
+            ImGui.Spacing();
+            ImGui.Separator();
+            ImGui.Spacing();
+            ImGui.Spacing();
+
+            if (ImGui.Button("Open Settings Folder"))
+            {
+                Util.Extensions.OpenFolder(api.PluginInterface.ConfigDirectory.FullName);
+                ImGuiUtil.AddNotification(NotificationType.Success, $"Settings exported");
+            }
+
+            ImGui.Spacing();
+            ImGui.Spacing();
+        }
+
+        ImGuiGroupPanel.EndGroupPanel();
+    }
+
+    private void RunSetDefaultPerformerFolderImGui()
+    {
+        fileDialogManager.OpenFolderDialog("Set Default Performer Folder", (result, filePath) =>
+        {
+            // PluginLog.Debug($"dialog result: {result}\n{string.Join("\n", filePath)}");
+            if (result)
+            {
+                MidiFileConfigManager.SetDefaultPerformerFolder(filePath);
+                MidiBard.SaveConfig();
+                IPCHandles.SyncAllSettings();
+                IPCHandles.UpdateDefaultPerformer();
+            }
+        }, MidiBard.config.defaultPerformerFolder);
+    }
+
+}
